requirements.txt
# 数据库驱动
mysql-connector-python>=8.0.0
psycopg2-binary>=2.8.0

# 开发依赖
pytest>=6.0.0